refactor: AppmapMiddleware should inherit from ABC
AppmapMiddleware is an abstract class, so it should inherit from
abc.ABC. This will take care of ensuring that subclasses implement its
abtract methods, removing the need to throw an exception at runtime.

In addition, these changes

  *  move Recorder to recorder.py
  *  move Recording to recording.py
  *  move importing code into Importer
