+++ appmap/__init__.py
@@ -2,7 +2,8 @@
 from ._implementation import generation  # noqa: F401
 from ._implementation.env import Env  # noqa: F401
 from ._implementation.labels import labels  # noqa: F401
-from ._implementation.recording import Recording, instrument_module  # noqa: F401
+from ._implementation.importer import instrument_module  # noqa: F401
+from ._implementation.recording import Recording  # noqa: F401
 
 try:
     from . import flask  # noqa: F401
