+++ appmap/_implementation/__init__.py
@@ -1,6 +1,6 @@
 from . import configuration
 from . import env as appmapenv
-from . import event, metadata, recording
+from . import event, importer, metadata, recorder
 from .py_version_check import check_py_version
 
 
@@ -8,8 +8,9 @@ def initialize(**kwargs):
     check_py_version()
     appmapenv.initialize(**kwargs)
     event.initialize()
-    recording.initialize()
-    configuration.initialize()  # needs to be initialized after recording
+    importer.initialize()
+    recorder.initialize()
+    configuration.initialize()  # needs to be initialized after recorder
     metadata.initialize()
