+++ appmap/http.py
@@ -9,7 +9,7 @@ from http.client import HTTPConnection
 from urllib.parse import parse_qs, urlsplit
 
 from ._implementation.event import HttpClientRequestEvent, HttpClientResponseEvent
-from ._implementation.recording import Recorder
+from ._implementation.recorder import Recorder
 from ._implementation.utils import patch_class, values_dict
